% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPxAGE_bias.R
\name{SNPxAGE_bias}
\alias{SNPxAGE_bias}
\title{Estimate bias from age-varying genetic effects in Mendelian randomization}
\usage{
SNPxAGE_bias(SNPxAGE_model_output, rep = 2, age_set = 65)
}
\arguments{
\item{SNPxAGE_model_output}{Output from \code{SNPxAGE_model}}

\item{rep}{Numerical value of the number of iterations to run the simulations}
}
\value{
A list of results
}
\description{
Takes output from \code{SNPxAGE_model} and estimates the bias due to
age-varying genetic effects under different exposure windows and at two
different ages
}
\examples{



}
