% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPxAGE_effect.R
\name{SNPxAGE_effect}
\alias{SNPxAGE_effect}
\title{Function to estimate the per-allele effect at each age}
\usage{
SNPxAGE_effect(SNPxAGE_output, reps = 10, ages = 40:70)
}
\description{
Function to estimate the per-allele effect at each age
}
