% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPxAGE_model.R
\name{SNPxAGE_model}
\alias{SNPxAGE_model}
\title{Flexibly modeling the age/phenotype association by genetic variant}
\usage{
SNPxAGE_model(data, SNP, phenotype, age, covars, k = 3, pred_ages, type = "cr")
}
\arguments{
\item{data}{A data frame containing the SNP, age and phenotype variables}

\item{SNP}{A character string with the name of the SNP variable in the data frame. This variable itself must be a factor variable containing at most 3 unique values indicating the number of copies}

\item{phenotype}{A character string with the name of the phenotype variable in the data frame}

\item{age}{A character string with the name of the age variable in the data frame}

\item{covars}{A character vector with the names of the variables to be adjusted for (e.g. principal components)}

\item{type}{A character string with the type splines to be used (see \code{mgcv} function)}

\item{knots}{A numeric value indicating the number of internal knots (default 3)}
}
\value{
A list including the model and parameters used. If pred_ages is supplied then the predictions for those ages is also returned.
}
\description{
Takes a genetic variant or single nucleotide polymorphism (SNP) and a
phenotype it's associated with and estimates how the relationship between the
two varies with age using splines.
}
\references{
TBA
}
\author{
Jeremy A Labrecque, \email{j.labrecque@erasmusmc.nl}
}
