% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPxAGE_plot.R
\name{SNPxAGE_plot}
\alias{SNPxAGE_plot}
\title{Function to plot absolute value of pheontype for each genotype}
\usage{
SNPxAGE_plot(SNPxAGE_model_output)
}
\description{
Function to plot absolute value of pheontype for each genotype
}
